requirements.txt
fal-client
requests~=2.32.3
bs4~=0.0.2
beautifulsoup4~=4.12.3
tqdm~=4.66.5
aiohttp~=3.10.9
python-dotenv~=1.0.1
pydub~=0.25.1
yt-dlp~=2024.7.16
youtube-transcript-api~=0.6.2
openai~=1.51.0
aiohttp
aiofiles
google-cloud-aiplatform
selenium
webdriver-manager

#https://x.com/i/broadcasts/1mrGmMNdNWBGy